requirements.txt
langchain-community==0.0.2
langchain-openai==0.0.2
openai==0.27.8
neo4j==5.11.0
faiss-cpu==1.7.4
biopython==1.81